feat:CRM合同创建

+++ src/main/java/com/bdaim/crm/erp/crm/entity/base/BaseCrmContract.java
@@ -210,12 +210,12 @@ public abstract class BaseCrmContract<M extends BaseCrmContract<M>> extends Mode
 	public Integer getContactsId() {
 		return getInt("contacts_id");
 	}
-	public void setCompanyUserId(Integer companyUserId) {
+	public void setCompanyUserId(Long companyUserId) {
 		set("company_user_id", companyUserId);
 	}
 
-	public Integer getCompanyUserId() {
-		return getInt("company_user_id");
+	public Long getCompanyUserId() {
+		return getLong("company_user_id");
 	}
 	public String getRemark() {
 		return getStr("remark");
